+++ prng_transaction.go
@@ -1,0 +1,341 @@
+package hedera
+
+/*-
+ *
+ * Hedera Go SDK
+ *
+ * Copyright (C) 2020 - 2022 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/hashgraph/hedera-protobufs-go/services"
+)
+
+type PrngTransaction struct {
+	Transaction
+	rang uint32
+}
+
+func NewPrngTransaction() *PrngTransaction {
+	transaction := PrngTransaction{
+		Transaction: _NewTransaction(),
+	}
+
+	transaction.SetMaxTransactionFee(NewHbar(5))
+
+	return &transaction
+}
+
+func _PrngTransactionFromProtobuf(transaction Transaction, pb *services.TransactionBody) *PrngTransaction {
+	return &PrngTransaction{
+		Transaction: transaction,
+		rang:        uint32(pb.GetPrng().GetRange()),
+	}
+}
+
+func (transaction *PrngTransaction) SetGrpcDeadline(deadline *time.Duration) *PrngTransaction {
+	transaction.Transaction.SetGrpcDeadline(deadline)
+	return transaction
+}
+
+// SetPayerAccountID Sets an optional id of the account to be charged the service fee for the scheduled transaction at
+// the consensus time that it executes (if ever); defaults to the ScheduleCreate payer if not
+// given
+func (transaction *PrngTransaction) SetRange(r uint32) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.rang = r
+
+	return transaction
+}
+
+func (transaction *PrngTransaction) GetRange() uint32 {
+	return transaction.rang
+}
+
+func (transaction *PrngTransaction) _Build() *services.TransactionBody {
+	body := &services.PrngTransactionBody{
+		Range: int32(transaction.rang),
+	}
+
+	return &services.TransactionBody{
+		TransactionFee:           transaction.transactionFee,
+		Memo:                     transaction.Transaction.memo,
+		TransactionValidDuration: _DurationToProtobuf(transaction.GetTransactionValidDuration()),
+		TransactionID:            transaction.transactionID._ToProtobuf(),
+		Data: &services.TransactionBody_Prng{
+			Prng: body,
+		},
+	}
+}
+
+func (transaction *PrngTransaction) _ConstructScheduleProtobuf() (*services.SchedulableTransactionBody, error) {
+	body := &services.PrngTransactionBody{
+		Range: int32(transaction.rang),
+	}
+
+	return &services.SchedulableTransactionBody{
+		TransactionFee: transaction.transactionFee,
+		Memo:           transaction.Transaction.memo,
+		Data: &services.SchedulableTransactionBody_Prng{
+			Prng: body,
+		},
+	}, nil
+}
+
+func _RngTransactionGetMethod(request interface{}, channel *_Channel) _Method {
+	return _Method{
+		transaction: channel._GetUtil().Prng,
+	}
+}
+
+func (transaction *PrngTransaction) IsFrozen() bool {
+	return transaction._IsFrozen()
+}
+
+// Sign uses the provided privateKey to sign the transaction.
+func (transaction *PrngTransaction) Sign(
+	privateKey PrivateKey,
+) *PrngTransaction {
+	return transaction.SignWith(privateKey.PublicKey(), privateKey.Sign)
+}
+
+func (transaction *PrngTransaction) SignWithOperator(
+	client *Client,
+) (*PrngTransaction, error) {
+	// If the transaction is not signed by the _Operator, we need
+	// to sign the transaction with the _Operator
+
+	if client == nil {
+		return nil, errNoClientProvided
+	} else if client.operator == nil {
+		return nil, errClientOperatorSigning
+	}
+	if !transaction.IsFrozen() {
+		_, err := transaction.FreezeWith(client)
+		if err != nil {
+			return transaction, err
+		}
+	}
+	return transaction.SignWith(client.operator.publicKey, client.operator.signer), nil
+}
+
+// SignWith executes the TransactionSigner and adds the resulting signature data to the Transaction's signature map
+// with the publicKey as the map key.
+func (transaction *PrngTransaction) SignWith(
+	publicKey PublicKey,
+	signer TransactionSigner,
+) *PrngTransaction {
+	if !transaction._KeyAlreadySigned(publicKey) {
+		transaction._SignWith(publicKey, signer)
+	}
+
+	return transaction
+}
+
+// Execute executes the Transaction with the provided client
+func (transaction *PrngTransaction) Execute(
+	client *Client,
+) (TransactionResponse, error) {
+	if client == nil {
+		return TransactionResponse{}, errNoClientProvided
+	}
+
+	if transaction.freezeError != nil {
+		return TransactionResponse{}, transaction.freezeError
+	}
+
+	if !transaction.IsFrozen() {
+		_, err := transaction.FreezeWith(client)
+		if err != nil {
+			return TransactionResponse{}, err
+		}
+	}
+
+	transactionID := transaction.transactionIDs._GetCurrent().(TransactionID)
+
+	if !client.GetOperatorAccountID()._IsZero() && client.GetOperatorAccountID()._Equals(*transactionID.AccountID) {
+		transaction.SignWith(
+			client.GetOperatorPublicKey(),
+			client.operator.signer,
+		)
+	}
+
+	resp, err := _Execute(
+		client,
+		&transaction.Transaction,
+		_TransactionShouldRetry,
+		_TransactionMakeRequest,
+		_TransactionAdvanceRequest,
+		_TransactionGetNodeAccountID,
+		_RngTransactionGetMethod,
+		_TransactionMapStatusError,
+		_TransactionMapResponse,
+		transaction._GetLogID(),
+		transaction.grpcDeadline,
+		transaction.maxBackoff,
+		transaction.minBackoff,
+		transaction.maxRetry,
+	)
+
+	if err != nil {
+		return TransactionResponse{
+			TransactionID: transaction.GetTransactionID(),
+			NodeID:        resp.(TransactionResponse).NodeID,
+		}, err
+	}
+
+	hash, err := transaction.GetTransactionHash()
+	if err != nil {
+		return TransactionResponse{}, err
+	}
+
+	return TransactionResponse{
+		TransactionID:          transaction.GetTransactionID(),
+		NodeID:                 resp.(TransactionResponse).NodeID,
+		Hash:                   hash,
+		ScheduledTransactionId: transaction.GetTransactionID(),
+	}, nil
+}
+
+func (transaction *PrngTransaction) Freeze() (*PrngTransaction, error) {
+	return transaction.FreezeWith(nil)
+}
+
+func (transaction *PrngTransaction) FreezeWith(client *Client) (*PrngTransaction, error) {
+	if transaction.IsFrozen() {
+		return transaction, nil
+	}
+	transaction._InitFee(client)
+	if err := transaction._InitTransactionID(client); err != nil {
+		return transaction, err
+	}
+	body := transaction._Build()
+
+	return transaction, _TransactionFreezeWith(&transaction.Transaction, client, body)
+}
+
+func (transaction *PrngTransaction) GetMaxTransactionFee() Hbar {
+	return transaction.Transaction.GetMaxTransactionFee()
+}
+
+// SetMaxTransactionFee sets the max transaction fee for this PrngTransaction.
+func (transaction *PrngTransaction) SetMaxTransactionFee(fee Hbar) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.Transaction.SetMaxTransactionFee(fee)
+	return transaction
+}
+
+// SetRegenerateTransactionID sets if transaction IDs should be regenerated when `TRANSACTION_EXPIRED` is received
+func (transaction *PrngTransaction) SetRegenerateTransactionID(regenerateTransactionID bool) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.Transaction.SetRegenerateTransactionID(regenerateTransactionID)
+	return transaction
+}
+
+// GetRegenerateTransactionID returns true if transaction ID regeneration is enabled.
+func (transaction *PrngTransaction) GetRegenerateTransactionID() bool {
+	return transaction.Transaction.GetRegenerateTransactionID()
+}
+
+func (transaction *PrngTransaction) GetTransactionMemo() string {
+	return transaction.Transaction.GetTransactionMemo()
+}
+
+// SetTransactionMemo sets the memo for this PrngTransaction.
+func (transaction *PrngTransaction) SetTransactionMemo(memo string) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.Transaction.SetTransactionMemo(memo)
+	return transaction
+}
+
+func (transaction *PrngTransaction) GetTransactionValidDuration() time.Duration {
+	return transaction.Transaction.GetTransactionValidDuration()
+}
+
+// SetTransactionValidDuration sets the valid duration for this PrngTransaction.
+func (transaction *PrngTransaction) SetTransactionValidDuration(duration time.Duration) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.Transaction.SetTransactionValidDuration(duration)
+	return transaction
+}
+
+func (transaction *PrngTransaction) GetTransactionID() TransactionID {
+	return transaction.Transaction.GetTransactionID()
+}
+
+// SetTransactionID sets the TransactionID for this PrngTransaction.
+func (transaction *PrngTransaction) SetTransactionID(transactionID TransactionID) *PrngTransaction {
+	transaction._RequireNotFrozen()
+
+	transaction.Transaction.SetTransactionID(transactionID)
+	return transaction
+}
+
+// SetNodeAccountID sets the _Node AccountID for this PrngTransaction.
+func (transaction *PrngTransaction) SetNodeAccountIDs(nodeID []AccountID) *PrngTransaction {
+	transaction._RequireNotFrozen()
+	transaction.Transaction.SetNodeAccountIDs(nodeID)
+	return transaction
+}
+
+func (transaction *PrngTransaction) SetMaxRetry(count int) *PrngTransaction {
+	transaction.Transaction.SetMaxRetry(count)
+	return transaction
+}
+
+func (transaction *PrngTransaction) SetMaxBackoff(max time.Duration) *PrngTransaction {
+	if max.Nanoseconds() < 0 {
+		panic("maxBackoff must be a positive duration")
+	} else if max.Nanoseconds() < transaction.minBackoff.Nanoseconds() {
+		panic("maxBackoff must be greater than or equal to minBackoff")
+	}
+	transaction.maxBackoff = &max
+	return transaction
+}
+
+func (transaction *PrngTransaction) GetMaxBackoff() time.Duration {
+	if transaction.maxBackoff != nil {
+		return *transaction.maxBackoff
+	}
+
+	return 8 * time.Second
+}
+
+func (transaction *PrngTransaction) SetMinBackoff(min time.Duration) *PrngTransaction {
+	if min.Nanoseconds() < 0 {
+		panic("minBackoff must be a positive duration")
+	} else if transaction.maxBackoff.Nanoseconds() < min.Nanoseconds() {
+		panic("minBackoff must be less than or equal to maxBackoff")
+	}
+	transaction.minBackoff = &min
+	return transaction
+}
+
+func (transaction *PrngTransaction) GetMinBackoff() time.Duration {
+	if transaction.minBackoff != nil {
+		return *transaction.minBackoff
+	}
+
+	return 250 * time.Millisecond
+}
+
+func (transaction *PrngTransaction) _GetLogID() string {
+	timestamp := transaction.transactionIDs._GetCurrent().(TransactionID).ValidStart
+	return fmt.Sprintf("RngTransaction:%d", timestamp.UnixNano())
+}