+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## v2.17.0
+
+### Added
+
+* `PrngThansaction`
+* `TransactionRecord.PrngBytes`
+* `TransactionRecord.PrngNumber`
+* `task` runner support
+
+## Deprecated
+
+* `ContractFunctionResult.ContractStateChanges` with no replacement.
+
+## v2.17.0-beta.1
+
+### Added
+
+* `PrngThansaction`
+* `TransactionRecord.PrngBytes`
+* `TransactionRecord.PrngNumber`
+
+## Deprecated
+
+* `ContractFunctionResult.ContractStateChanges` with no replacement.
+
 ## v2.16.1
 
 ### Added